requirements.txt
confluent-kafka>=2.3.0
tweepy==4.14.0
psycopg2-binary==2.9.10
textblob==0.17.1
vaderSentiment==3.3.2
python-dotenv==1.0.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit>=1.29.0
plotly>=5.18.0
altair>=5.2.0
streamlit-autorefresh==0.0.1